Updated $$channel$$.js to support oneOf: messages
Modified the validateMessage function to validate the payload and if
valid, increment a counter of validated messages. Ignore validation
errors. Throw an exception only if nValidated exceeds 1 (no need to
try validating more). router function tries to validate the payload
against all of the messages. If exactly one matches, then we proceed,
else we stop.