Updated `$$channel$$.js` (and `lib/message-validator.js`) to support `oneOf:` messages
Modified the `validateMessage` function to validate the payload and if
valid, increment and return a counter of validated messages. Ignore
validation exceptions. Throw an exception only if `nValidated > 1` (no
need to try validating more - fail fast). `router` functions (publish
or subscribe) try to validate the payload against all of the messages.
If exactly one matches, then we proceed, else we stop.